Criação do main do projeto

+++ src/main/java/br/senac/eco2you/modelo/entidade/conquista/Conquista.java
@@ -42,7 +42,7 @@ public class Conquista implements Serializable {
 	public Conquista() {
 	}
 	
-	public Conquista(double pontosDeposito, Deposito deposito, String descricao, String nome) {
+	public Conquista(float pontosDeposito, Deposito deposito, String descricao, String nome) {
 		setPontosDeposito(pontosDeposito);
 		setDeposito(deposito);
 		setDescricao(descricao);
